Novo App Vendas/backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-jose==3.3.0
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
drf-yasg==1.21.7
